# public/openapi.yaml
openapi: 3.0.0
info:
  title: SNF Forms Catalog API
  description: API for accessing the SNF Forms catalog, including form details and search functionality.
  version: 1.0.0

paths:
  /api/forms:
    get:
      summary: List all forms
      description: Retrieve a list of all forms in the catalog, optionally filtered by category or search query.
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter forms by exact category match (case-insensitive).
        - in: query
          name: query
          schema:
            type: string
          description: Search forms by ID or description (case-insensitive).
      responses:
        '200':
          description: A list of forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Form'
        '500':
          description: Server error

  /api/forms/{formId}:
    get:
      summary: Get a form by ID
      description: Retrieve a single form by its unique ID.
      parameters:
        - in: path
          name: formId
          required: true
          schema:
            type: string
          description: The unique identifier of the form (e.g., SNF-24).
      responses:
        '200':
          description: The requested form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '404':
          description: Form not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Form not found
        '500':
          description: Server error

  /api/search:
    get:
      summary: Search forms
      description: Search the forms catalog with pagination support.
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Search term for ID, description, or category.
        - in: query
          name: category
          schema:
            type: string
          description: Filter by exact category match.
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Maximum number of results to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of results to skip.
      responses:
        '200':
          description: Search results with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Form'
                  total:
                    type: integer
                    description: Total number of matching forms.
                  limit:
                    type: integer
                    description: The limit used for this request.
                  offset:
                    type: integer
                    description: The offset used for this request.
        '500':
          description: Server error

components:
  schemas:
    Form:
      type: object
      properties:
        formId:
          type: string
          description: Unique identifier for the form.
          example: SNF-24
        category:
          type: string
          description: Product category.
          example: Nursing
        description:
          type: string
          description: Detailed description or name of the form.
          example: 24 Hour Nursing Report
        size:
          type: string
          description: Physical dimensions of the form.
          example: 8-1/2 x 14
        paper:
          type: string
          description: Paper type or weight.
          example: 2 part NCR
        color:
          type: string
          description: Ink colors.
          example: White/Yellow
        sides:
          type: string
          description: Number of printed sides.
          example: "1"
        unit:
          type: string
          description: Packaging unit.
          example: S/100
        file0:
          type: string
          description: URL or reference to the primary image.
        file1:
          type: string
          description: URL or reference to the secondary image.
        pdf0:
          type: string
          description: URL or reference to the PDF version.